refact: 비밀 캡슐 쿼리 분리
